# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RegisAI Compliance Platform API
  version: 0.1.0
  description: |
    AI-powered compliance analysis platform. Upload compliance documentation,
    run gap analysis against regulatory frameworks, draft remediation policy
    language, and monitor regulatory changes.

    **Authentication:** All endpoints require a valid Supabase session cookie
    (set via /login or /auth/callback). Unauthenticated requests return 401.

    **Data isolation:** Row-Level Security enforces per-user data isolation.
    Users can only access their own documents, audits, and findings.
  contact:
    name: RegisAI Team
    url: https://github.com/bitcoinow/RegisAI
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://regis-ai-nine.vercel.app
    description: Production (Vercel)
...
paths:
...
  /api/findings/{id}/status:
    patch:
      tags: [Findings]
      summary: Update finding status
      description: |
        Transitions a finding's status. Moving away from 'open' records the reviewer
        identity, timestamp, and optional review note as an audit trail.
      operationId: updateFindingStatus
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/FindingStatus"
                review_note:
                  type: string
                  description: Optional rationale for the status change
            examples:
              resolve:
                summary: Mark finding as resolved
                value:
                  status: resolved
                  review_note: "Policy updated in section 4.2"
              accept_risk:
                summary: Accept the risk
                value:
                  status: risk_accepted
                  review_note: "Low materiality — board accepted residual risk"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingStatusResponse"
        "400":
          description: Invalid status value
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Finding not found or not owned by user
...
  /api/monitoring/feed:
    get:
      tags: [Monitoring]
      summary: Get regulatory update feed
      description: |
        Returns stored regulatory updates from FCA, PRA, Bank of England, SEC,
        FINRA, ESMA, and EBA. Ordered by publication date descending. Max 500 items.
      operationId: getMonitoringFeed
      responses:
        "200":
          description: Regulatory updates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/monitoring/refresh:
    post:
      tags: [Monitoring]
      summary: Refresh regulatory feeds
      description: |
        Fetches latest updates from all regulatory feed sources (Federal Register API
        for US; RSS feeds for FCA, PRA, ESMA, EBA). Upserts into the database,
        deduplicating by URL.
      operationId: refreshMonitoringFeeds
      responses:
        "200":
          description: Feed refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted:
                    type: integer
                    description: Number of updates inserted/updated
                  parsed:
                    type: integer
                    description: Number of updates parsed from sources
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/monitoring/digest:
    post:
      tags: [Monitoring]
      summary: Send weekly regulatory digest email
      description: Sends a regulatory digest email via Resend to the authenticated user.
      operationId: sendDigest
      responses:
        "200":
          description: Digest sent
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/profile:
    patch:
      tags: [Profile]
      summary: Update firm profile
      description: |
        Upserts the authenticated user's firm profile. Used during onboarding
        and in settings. All fields are required.
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firm_name, firm_type, aum_range, regulator]
              properties:
                firm_name:
                  type: string
                  example: "Meridian Wealth Advisers Ltd"
                firm_type:
                  $ref: "#/components/schemas/FirmType"
                aum_range:
                  type: string
                  example: "$50M-$100M"
                regulator:
                  $ref: "#/components/schemas/Regulator"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          description: Invalid profile fields
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    Jurisdiction:
      type: string
      enum: [US, EU, UK]
      default: US
...
    FindingStatus:
      type: string
      enum: [open, in_progress, resolved, risk_accepted]
    FirmType:
      type: string
      enum: [RIA, Fintech, Insurance, Bank]
    Regulator:
      type: string
      enum: [FINRA, SEC, State, Multiple]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
...
    FindingStatusResponse:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/FindingStatus"
        reviewed_by:
          type: string
          nullable: true
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        review_note:
          type: string
          nullable: true
        reviewer_label:
          type: string
          nullable: true
          description: Display name/email of the reviewer
...
    FeedResponse:
      type: object
      properties:
        updates:
          type: array
          items:
            $ref: "#/components/schemas/RegulatoryUpdate"
    RegulatoryUpdate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        regulator:
          type: string
        jurisdiction:
          $ref: "#/components/schemas/Jurisdiction"
        title:
          type: string
        summary:
          type: string
          nullable: true
        url:
          type: string
          format: uri
        published_at:
          type: string
          format: date-time
          nullable: true
        relevance_score:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        affected_rules:
          type: array
          items:
            type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firm_name:
          type: string
          nullable: true
        firm_type:
          $ref: "#/components/schemas/FirmType"
        aum_range:
          type: string
          nullable: true
        regulator:
          $ref: "#/components/schemas/Regulator"
        plan:
          type: string
        created_at:
          type: string
          format: date-time